GRPO+LMDeploy 0.7 (#3277)
